examples/zone-printer: Rewrite instructions
I rewrote the tutorial to make it follow our conventions and tone. It may
seem a bit more boring and robotic, but I think this makes it easier to
follow.

I renamed clusters to cluster-1 and cluster-2, and got rid of the asia-* one.

Also some simplifications around the kubeconfig file names and reduced the
amount of information people need to know about the kubeconfig files in general
so I added .yaml extension to these files.

Also added command outputs to some of the commands listed, to give an idea of
what it would look like.

I am planning on simplifying the "how does it work?" part by rewriting the
sample without using nginx (because nginx is often confused as it is a load
balancer itself).

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>