IngressSyncer: Fix a bug in DeleteIngress and add more unit tests.
ingresssyncer.go: When DeleteIngress delete an ingress, it checks the error to see if it was "not found". But the bug was that the logic was checking the wrong error variable, so it would never go down the ingress "not found" path.

A new unit test now verifies this logic.